manta-matcher/requirements.txt
fastapi
uvicorn
pillow
python-multipart
aiofiles
tqdm
numpy
openai
httpx
supabase
pgvector

